OPENEUROPA-1379: Rework implementation and tests for DateTimeItem.
